ai-service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.9.0
numpy>=1.26.4
pandas>=2.0.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0